shared: rename lib modules using lowercase
