+++ shared/lib/network.js
