+++ templates/democracy/test/democracy.js
@@ -1,9 +1,9 @@
 const { hash: namehash } = require('eth-ens-namehash')
 const { APP_IDS } = require('@aragon/templates-shared/helpers/apps')
 const { isGanache } = require('@aragon/templates-shared/helpers/node')(web3)
-const { isLocalNetwork } = require('@aragon/templates-shared/lib/Network')(web3)
+const { isLocalNetwork } = require('@aragon/templates-shared/lib/network')(web3)
 const { encodeCallScript } = require('@aragon/test-helpers/evmScript')
-const { deployedAddresses } = require('@aragon/templates-shared/lib/ArappFile')(web3)
+const { deployedAddresses } = require('@aragon/templates-shared/lib/arapp-file')(web3)
 
 const timeTravel = require('@aragon/test-helpers/timeTravel')(web3)
 const getBalance = require('@aragon/test-helpers/balance')(web3)
